# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Network Diagnostics Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /diagnostics/traceroute:
    post:
      operationId: runTraceroute
      tags: [diagnostics]
      summary: Run traceroute diagnostic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TracerouteInput"
      responses:
        "200":
          description: Traceroute results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticRun"
  /diagnostics/dns:
    post:
      operationId: runDns
      tags: [diagnostics]
      summary: Run DNS lookup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DnsInput"
      responses:
        "200":
          description: DNS lookup results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticRun"
  /diagnostics/port-check:
    post:
      operationId: runPortCheck
      tags: [diagnostics]
      summary: Check TCP port reachability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortCheckInput"
      responses:
        "200":
          description: Port check results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticRun"
...
  /export/run/{id}/json:
    get:
      operationId: exportRunJson
      tags: [export]
      summary: Export a diagnostic run as JSON
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: JSON export
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticRun"
  /export/run/{id}/csv:
    get:
      operationId: exportRunCsv
      tags: [export]
      summary: Export a diagnostic run as CSV
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
...
components:
  schemas:
...
    TracerouteInput:
      type: object
      properties:
        target:
          type: string
          minLength: 1
        maxHops:
          type: integer
          minimum: 1
          maximum: 30
        timeout:
          type: integer
          minimum: 1
          maximum: 60
      required:
        - target
    DnsInput:
      type: object
      properties:
        target:
          type: string
          minLength: 1
        recordType:
          type: string
          enum: [A, AAAA, MX, NS, TXT, CNAME, PTR, SOA]
        server:
          type: ["string", "null"]
      required:
        - target
    PortCheckInput:
      type: object
      properties:
        target:
          type: string
          minLength: 1
        port:
          type: integer
          minimum: 1
          maximum: 65535
        timeout:
          type: integer
          minimum: 1
          maximum: 30
      required:
        - target
        - port
    DiagnosticRun:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        target:
          type: string
        status:
          type: string
          enum: [success, failure, partial, running]
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]
        durationMs:
          type: ["integer", "null"]
        demoMode:
          type: boolean
        rawOutput:
          type: ["string", "null"]
        parsedResult:
          $ref: "#/components/schemas/ParsedResult"
        metrics:
          $ref: "#/components/schemas/RunMetrics"
        hops:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/RouteHop"
        dnsRecords:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/DnsRecord"
        error:
          type: ["string", "null"]
        parseWarnings:
          type: array
          items:
            type: string
      required:
        - id
        - type
        - target
        - status
        - startedAt
        - demoMode
        - parseWarnings
    ParsedResult:
      type: object
      properties:
        summary:
          type: string
        success:
          type: boolean
        details:
          type: object
      required:
        - summary
        - success
    RunMetrics:
      type: object
      properties:
        latencyMin:
          type: ["number", "null"]
        latencyMax:
          type: ["number", "null"]
        latencyAvg:
          type: ["number", "null"]
        packetLoss:
          type: ["number", "null"]
        jitter:
          type: ["number", "null"]
        hopCount:
          type: ["integer", "null"]
        packetsTransmitted:
          type: ["integer", "null"]
        packetsReceived:
          type: ["integer", "null"]
    RouteHop:
      type: object
      properties:
        hop:
          type: integer
        host:
          type: ["string", "null"]
        ip:
          type: ["string", "null"]
        rtt1:
          type: ["number", "null"]
        rtt2:
          type: ["number", "null"]
        rtt3:
          type: ["number", "null"]
        timeout:
          type: boolean
      required:
        - hop
        - timeout
    DnsRecord:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        value:
          type: string
        ttl:
          type: ["integer", "null"]
      required:
        - type
        - name
        - value
...