# lib/api-zod/src/generated/types/parsedResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Network Diagnostics Dashboard API
 * OpenAPI spec version: 1.0.0
 */
import type { ParsedResultDetails } from './parsedResultDetails';

export interface ParsedResult {
  summary: string;
  success: boolean;
  details?: ParsedResultDetails;
}