# lib/api-zod/src/generated/types/routeHop.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Network Diagnostics Dashboard API
 * OpenAPI spec version: 1.0.0
 */

export interface RouteHop {
  hop: number;
  /** @nullable */
  host?: string | null;
  /** @nullable */
  ip?: string | null;
  /** @nullable */
  rtt1?: number | null;
  /** @nullable */
  rtt2?: number | null;
  /** @nullable */
  rtt3?: number | null;
  timeout: boolean;
}