# lib/api-zod/src/generated/types/tracerouteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Network Diagnostics Dashboard API
 * OpenAPI spec version: 1.0.0
 */

export interface TracerouteInput {
  /** @minLength 1 */
  target: string;
  /**
     * @minimum 1
     * @maximum 30
     */
  maxHops?: number;
  /**
     * @minimum 1
     * @maximum 60
     */
  timeout?: number;
}